Require java 8 as minimum version for building & running Hazelcast (#14385)
Require java 8 as minimum version for building & running Hazelcast

Adds infrastructure to support JDK 8 internals and JDK 6 build
(see PlatformSpecific class)

+++ hazelcast/src/test/java/com/hazelcast/test/starter/answer/test/AnswerTest.java
@@ -217,7 +217,8 @@ public class AnswerTest extends HazelcastTestSupport {
         Data data = serializationService.toData(original);
         assertNotNull("data should not be null", data);
         assertFalse("data should be no proxy class", isProxyClass(data.getClass()));
-        assertEquals("toObject() should return original value", original, serializationService.toObject(data));
+        assertEquals("toObject() should return original value", original,
+                ((Integer) serializationService.toObject(data)).intValue());
 
         SerializationService localSerializationService = new DefaultSerializationServiceBuilder().build();
         Data localData = localSerializationService.toData(original);
